# ingest-api-spec.yaml
openapi: 3.0.0
info:
  description: |
    A restful web service exposing calls needed for the Ingest UI React application.
  version: 2.0.20
  title: HuBMAP Ingest API
  contact:
    name: HuBMAP Help Desk
    email: help@hubmapconsortium.org
  license:
    name: MIT License
    url: 'https://github.com/hubmapconsortium/ingest-api/blob/main/license.txt'
...
paths:
...
  /metadata/validate:
    post:
      summary: "Validate uploaded tsv file"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pathname:
                  type: string
                entity_type:
                  type: string
                sub_type:
                  type: string
                validate_uuids:
                  type: string
                tsv_row:
                  type: string
      responses:
        '200':
          description: The tsv file was successfully validated.
        '400':
          description: Missing input
        '406':
          description: The tsv file was not acceptable
        '500':
          description: Internal error
servers:
  - url: "https://uuid.hubmapconsortium.org/dataingest"
  - url: "http://uuid.hubmapconsortium.org/dataingest"
components:
  securitySchemes:
    globus_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: "https://auth.globus.org/v2/oauth2/authorize"
          scopes:
            "write:dataset": modify datasets
            "read:dataset": read datasets
    api_key:
      type: apiKey
      name: api_key
      in: header
...